Refactor sign up page
